chore: changed file structure in playground a bit, updated dependency
